Improve hourly forecast trend: midnight-to-midnight range, current-hour centering, day sync
- Hourly trend on the main screen and details page now always spans from
  midnight onward (weather.hourlyTrendForecast/hourlyTrendCurrentIndex),
  across multiple days, instead of a rolling 24h-from-now window.
- Highlight and horizontally center the current hour when the hourly card
  opens, both on the main screen (DailyViewHolder) and the details
  "Conditions" page (DetailsConditions) -- centering on that specific day's
  first hour instead when viewing a day other than today.
- Details page: the day-tab row now follows along as the hourly strip is
  scrolled across a midnight boundary, without changing the rest of the page.
- Remove the wind-speed number under the wind arrow (main hourly wind card)
  and the UV index row (combined temperature/conditions card) -- redundant
  clutter, not part of the original design intent.
- Fix an hourly-position mismatch on the details page introduced by the
  above (tap-to-select an hour used a differently-scoped list than the
  adapter's own data).
- Minor build-warning cleanup: deprecated clipboard/Firebase/Compose APIs,
  a redundant conversion, unchecked-cast suppressions, a smart-cast fix.
- Add docs/FUNCTIONEEL_DOCUMENT.md: a from-source-verified functional
  overview of the app, correcting several stale points in existing docs.

+++ app/src/main/kotlin/com/liveweatherwallpaperapp/wallpaper/WallpaperSceneState.kt
@@ -575,7 +575,7 @@ object WallpaperSceneStateFactory {
         } else {
             0.45f + (beaufort - 8f) / 4f * 0.45f
         }.coerceIn(0f, 0.90f)
-        val angleRadians = Math.toRadians((tiltFraction * 90.0).toDouble())
+        val angleRadians = Math.toRadians(tiltFraction * 90.0)
         val magnitude = tan(angleRadians).toFloat()
         val direction = normalizeDegrees(windDirectionDegrees)
         val sign = if (direction == null || cos(Math.toRadians(direction.toDouble())) >= 0.0) 1f else -1f
